Move the menus for object/artifact creation and the projection graphic effects to ui-wizard.c from wiz-debug.c.
