+++ src/Makefile.src
@@ -208,6 +208,7 @@ ANGFILES = \
 	ui-target.o \
 	ui-term.o \
 	ui-visuals.o \
+	ui-wizard.o \
 	wiz-debug.o \
 	wiz-spoil.o \
 	wiz-stats.o \
