+++ src/ui-wizard.h
@@ -1,0 +1,22 @@
+/**
+ * \file ui-wizard.h
+ * \brief Declarations for menus related to debug commands
+ *
+ * This work is free software; you can redistribute it and/or modify it
+ * under the terms of either:
+ *
+ * a) the GNU General Public License as published by the Free Software
+ *    Foundation, version 2, or
+ *
+ * b) the "Angband licence":
+ *    This software may be copied and distributed for educational, research,
+ *    and not for profit purposes provided that this copyright and statement
+ *    are included in all such copies.  Other copyrights may also apply.
+ */
+#ifndef INCLUDED_UI_WIZARD_H
+#define INCLUDED_UI_WIZARD_H
+
+void wiz_create_item(bool art);
+void wiz_proj_demo(void);
+
+#endif /* INCLUDED_UI_WIZARD_H */
